login and register implemented in frontent, error popup implemented in failed login
