+++ frontend/src/components/AuthProvider.jsx
@@ -3,9 +3,9 @@ import React, { createContext, useState, useContext } from 'react';
 const AuthContext = createContext();
 
 export const AuthProvider = ({ children }) => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [isLoggedIn, setIsLoggedIn] = useState(true);
 
-  const login = () => setIsLoggedIn(true);
+  const login = () => setIsLoggedIn(false);
   const logout = () => setIsLoggedIn(false);
 
   return (
